修改文档

+++ docs/.vuepress/theme/Layout.vue
@@ -55,7 +55,6 @@
   </div>
 </template>
 <script>
-import Dialog from "../../../components/lib/dialog/index.js";
 export default {
   data() {
     return {
@@ -80,13 +79,12 @@ export default {
   created() {},
   mounted() {
     // console.log(this.$site.base,'this.$sitethis.$site')
-
     if (this.$page.path == "/") {
       window.location.href =
         "https://he-guang-wen.github.io/water-ui/docs-dist/guide/quickstart.html";
       return;
     }
-        this.isPc();
+    this.isPc();
     window.addEventListener(
       "message",
       (e) => {
@@ -179,19 +177,14 @@ export default {
       sessionStorage.setItem("asideTop", this.$refs.aside.scrollTop);
     },
     isPc() {
-      if(sessionStorage.getItem('hideDialog')) return false
+      if (sessionStorage.getItem("hideDialog")) return false;
       if (/Android|webOS|iPhone|iPod|BlackBerry/i.test(navigator.userAgent)) {
-        Dialog({
-          title: "提示",
-          content: "检测到当前为移动端，是否切换到组件演示地址",
-        })
-          .then((res) => {
-            window.location.href =
-              "https://he-guang-wen.github.io/water-ui/dist/";
-          })
-          .catch((err) => {
-            sessionStorage.setItem('hideDialog',true)
-          });
+        if (confirm("检测当前为移动端，是否切换到演示地址")) {
+          window.location.href =
+            "https://he-guang-wen.github.io/water-ui/dist/";
+        } else {
+          sessionStorage.setItem("hideDialog", true);
+        }
         return false;
       } else {
         return true;
